Clear directories only with CleanBuild=true option

+++ vars/rpr_hybrid_ue_pipeline.groovy
@@ -94,8 +94,10 @@ def executeBuildWindows(String projectName, Map options) {
     String targetDir = projectsInfo[projectName]["targetDir"]
     String svnRepoName = projectsInfo[projectName]["svnRepoName"]
 
-    bat("if exist \"${targetDir}\" rmdir /Q /S ${targetDir}")
-    bat("if exist \"RPRHybrid-UE\" rmdir /Q /S RPRHybrid-UE")
+    if (options.cleanBuild) {
+        bat("if exist \"${targetDir}\" rmdir /Q /S ${targetDir}")
+        bat("if exist \"RPRHybrid-UE\" rmdir /Q /S RPRHybrid-UE")
+    }
 
     utils.removeFile(this, "Windows", "*.log")
